requirements.txt
captum
datasets
editdistance
empath
gensim
huggingface_hub
langchain
liwc
matplotlib
networkx
numpy
openai
pandas
pyannote.core
pyannote.database
pygamma_agreement
pynvml
regex
scikit_learn
seaborn
sortedcontainers
spacy
torch
transformers
xgboost
langchain
tabulate